✨ (bleu-rouge) Completed documentation, refactored LCS methods so that they are outside Rouge-L
